feat: remove unused react-popper from code
