+++ src/components/Feedback/controls/FeedbackFormPopup.tsx
@@ -13,7 +13,8 @@ import React, {
 import {Button, Checkbox, Popup, TextArea, useDirection} from '@gravity-ui/uikit';
 import block from 'bem-cn-lite';
 
-import {PopperPosition, useTranslation} from '../../../hooks';
+import {PopperPosition} from '../../../constants';
+import {useTranslation} from '../../../hooks';
 import {getPopupPosition} from '../../../utils';
 import {ControlsLayoutContext} from '../../Controls/ControlsLayout';
 import {FeedbackView} from '../Feedback';
