+++ src/components/Feedback/controls/SuccessPopup.tsx
@@ -4,7 +4,8 @@ import React, {memo, useContext, useMemo} from 'react';
 import {Popup, useDirection} from '@gravity-ui/uikit';
 import block from 'bem-cn-lite';
 
-import {PopperPosition, useTranslation} from '../../../hooks';
+import {PopperPosition} from '../../../constants';
+import {useTranslation} from '../../../hooks';
 import {getPopupPosition} from '../../../utils';
 import {ControlsLayoutContext} from '../../Controls/ControlsLayout';
 import {FeedbackView} from '../Feedback';
