+++ src/constants.ts
@@ -29,3 +29,18 @@ export const LEGACY_LANG_ITEMS = [
     {value: Lang.He, text: 'Hebrew'},
 ];
 export const SUPPORTED_LANGS = Object.keys(langs);
+
+export enum PopperPosition {
+    BOTTOM_START = 'bottom-start',
+    BOTTOM = 'bottom',
+    BOTTOM_END = 'bottom-end',
+    TOP_START = 'top-start',
+    TOP = 'top',
+    TOP_END = 'top-end',
+    RIGHT_START = 'right-start',
+    RIGHT = 'right',
+    RIGHT_END = 'right-end',
+    LEFT_START = 'left-start',
+    LEFT = 'left',
+    LEFT_END = 'left-end',
+}
