+++ src/hooks/index.ts
@@ -1,4 +1,3 @@
-export * from './usePopper';
 export * from './useTimer';
 export * from './useTranslation';
 export * from './usePopupState';
